# Conflicts:
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/actualworkinghours/ActualWorkingTimeOfDaily.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/actualworkinghours/AttendanceTimeOfDailyPerformance.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/actualworkinghours/ConstraintTime.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/actualworkinghours/SubHolOccurrenceInfo.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/actualworkinghours/WorkScheduleTimeOfDaily.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/breakorgoout/BreakTimeSheet.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/breakorgoout/OutingTimeSheet.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/dailyattendanceitem/DailyAttendanceItem.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/dailyattendanceitem/repository/DailyAttendanceItemNameDomainServiceRecImpl.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/divergencetimeofdaily/DivergenceTimeOfDaily.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/premiumtime/PremiumTimeOfDailyPerformance.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/shorttimework/ShortWorkingTimeSheet.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/stamp/StampItem.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/worktime/TimeActualStamp.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/worktime/TimeLeavingOfDailyPerformance.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/worktime/TimeLeavingWork.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/worktime/WorkStamp.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/worktime/repository/TimeLeavingOfDailyPerformanceRepository.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.infra/src/main/java/nts/uk/ctx/at/record/infra/entity/daily/overtimework/KrcdtDayOvertimework.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.infra/src/main/java/nts/uk/ctx/at/record/infra/entity/dailyattendanceitem/KshstControlOfAttendanceItems.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.infra/src/main/java/nts/uk/ctx/at/record/infra/repository/workrecord/worktime/JpaTimeLeavingOfDailyPerformanceRepository.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/common/time/AttendanceTime.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/employmentrule/hourlate/overtime/overtimeframe/OvertimeFrame.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/statutory/worktime/company/CompanyWtSetting.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/worktime/WorkTime.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/worktime/WorkTimeDailyAtr.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/worktime/WorkTimeDivision.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/worktime/WorkTimeMethodSet.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/worktime/WorkTimeRepository.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/worktime/fixedworkset/FixOffdayWorkTime.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/worktime/fixedworkset/FixRestTime.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/worktime/fixedworkset/FixRestTimeSetting.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/worktime/fixedworkset/FixWeekdayWorkTime.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/worktime/fixedworkset/FixedWorkSetting.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/worktime/flexworkset/FlexWorkSetting.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/worktime/fluidworkset/FluOffdayWorkTime.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/worktime/fluidworkset/FluRestTime.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/worktime/fluidworkset/FluRestTimeGroup.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/worktime/fluidworkset/FluRestTimeSetting.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/worktime/fluidworkset/FluWeekdayWorkTime.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/worktime/fluidworkset/FluidTimeSetting.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/worktime/fluidworkset/FluidWorkSetting.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/worktime/fluidworkset/TimeRoundingSetting.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/worktime/fluidworkset/Unit.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/worktimeset/PrescribedTimezoneSetting.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/worktimeset/Timezone.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/worktimeset/WorkTimeSet.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/worktimeset/common/CommonRestSetting.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/worktimeset/common/FlowFixedRestCalcMethod.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/worktimeset/common/FlowFixedRestSet.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/worktimeset/common/FlowRestClockCalcMethod.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/worktimeset/common/FlowWorkRestSetting.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/worktimeset/common/FlowWorkRestSettingDetail.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/worktimeset/common/RestTimeOfficeWorkCalcMethod.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/worktype/DailyWork.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.infra/src/main/java/nts/uk/ctx/at/shared/infra/repository/worktimeset/JpaWorkTimeSetRepository.java

+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.infra/src/main/java/nts/uk/ctx/at/record/infra/entity/optitem/KrcstCalcResultRange.java
@@ -45,27 +45,27 @@ public class KrcstCalcResultRange extends UkJpaEntity implements Serializable {
 
 	/** The upper time range. */
 	@Column(name = "UPPER_TIME_RANGE")
-	private int upperTimeRange;
+	private Integer upperTimeRange;
 
 	/** The lower time range. */
 	@Column(name = "LOWER_TIME_RANGE")
-	private int lowerTimeRange;
+	private Integer lowerTimeRange;
 
 	/** The upper number range. */
 	@Column(name = "UPPER_NUMBER_RANGE")
-	private int upperNumberRange;
+	private Double upperNumberRange;
 
 	/** The lower number range. */
 	@Column(name = "LOWER_NUMBER_RANGE")
-	private int lowerNumberRange;
+	private Double lowerNumberRange;
 
 	/** The upper amount range. */
 	@Column(name = "UPPER_AMOUNT_RANGE")
-	private int upperAmountRange;
+	private Integer upperAmountRange;
 
 	/** The lower amount range. */
 	@Column(name = "LOWER_AMOUNT_RANGE")
-	private int lowerAmountRange;
+	private Integer lowerAmountRange;
 
 	@OneToOne(optional = true, cascade = CascadeType.ALL)
 	@PrimaryKeyJoinColumns({ @PrimaryKeyJoinColumn(name = "CID", referencedColumnName = "CID"),
